Algorithm-specific code was moved to nested types
