--- src/Intelliflo.SDK.Security/Utils/IHashCalculator.cs
@@ -1,8 +1,0 @@
-﻿namespace Intelliflo.SDK.Security.Utils
-{
-    public interface IHashCalculator
-    {
-        string GetStringToSignHash(string value, string secret);
-        string GetCanonicalRequestHash(string canonicalRequest);
-    }
-}
